+++ Procfile
@@ -1,1 +1,2 @@
-worker: node bot.js
+worker: node glob.js
+worker: node zup.js
